pyproject.toml
[project]
name = "pydiscogstoqrfactory"
version = "0.1.0"
description = "Convert Discogs collection to QR Factory 3 CSV files for printing QR code stickers"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "flask>=3.0",
    "flask-session>=0.8",
    "flask-sqlalchemy>=3.1",
    "python3-discogs-client>=2.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydiscogstoqrfactory"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
